New `rivermigrate` Go API for running migrations from code + target version support (#67)
* New `rivermigrate` Go API for running migrations from code + target version support

Here, as requested by several users, add a Go API that's able to run
migrations as an alternative to migrating via the CLI. We add a new
package `rivermigrate` to mirror the conventions of `riverdriver` and
`rivertest`.

`rivermigrate` is largely just a refactor of the existing internal
package `internal/dbmigrate` with a couple important tweaks:

* We give it the ability to take a driver the same way as we do for
  `river.Client` and `rivertest`.

* To mirror client and `rivertest` we add a transaction-specific variant
  that takes a `TTx` (`MigrateTx`).

* Because we now have the non-transaction and transaction variants, I
  take away the `Down`/`Up` function distinctions, and make up/down a
  parameter to `Migrate`/`MigrateTx` instead (so we only need two
  functions instead of four).

* I move to using an `embed.FS` instead of specific embedded strings for
  each version and the file system is walked to get all migrations
  within it. This will make adding new migrations in the future easier
  and more foolproof.

I also added a new option for `TargetVersion` so that it's possible to
do a limited migration to a specific version instead of going all the
way up or down. This is a pretty standard feature in most migration
frameworks.

This still can't support Goose because it takes an `sql.DB`, but I think
we can get there by adding a minimal driver for the `sql` package that
can run migrations and nothing else. More of this to be explored later.

* Migrate only one step down by default + `TargetVersion` set to -1 for unlimited

A few tweaks to the originally proposed migration API:

* When migrating down, defaults to a maximum of one step by default.
  This is meant as a safety feature so that it's harder to remove the
  River schema completely by accident.

* When migrating down, `TargetVersion` can be set to `-1` explicitly to
  move an unlimited number of steps. This adds a way to remove the River
  schema completely by migrating to the equivalent of version zero (we
  can't use `0` for this because that's the default value of `int`).