Preparations for 0.0.13 release (#108)
Here, tee up the necessary changes for an upcoming 0.0.13 release that
I'll cut immediately after this is merged.

We update each submodule's `go.mod` to point to 0.0.13, which doesn't
actually exist yet, but it will very soon. This will leave the repo very
temporarily in an invalid state, but I tried the same change out in a
test project and as soon as the Go module proxy's cache busts, it works.

We comment out all the `replace`'s in `./cmd/river` so that it continues
to be installable with `go install ...@latest`. The `replace`'s are
still in there but commented out so that we can test future CLI changes
without too much trouble.

Finally, update the changelog to add the new section for 0.0.13.