Fix module references for River CLI (#110)
This one follows up #108 with a fix wherein we add concrete `go.sum`
entries for references to other River modules. This is necessary because
the CLI module no longer has `replace` directives.

This one was my mistake. As in #108, _most_ of the modules in River will
be releasable simultaneously, but the River CLI is special because it
doesn't have `replace` directives so it can stay `go install`able.

In the future, we'll probably set this up such that most of the modules
can be released together, but if the River CLI needs a new River
dependency version, we'll have to make sure to release the top-level
River packages first, then follow up the CLI with its own update with
the version bump.