Add script that updates River versions in `go.mod` files for release
Here, add a helper script whose job it is to during a release, go into
each `go.mod` submodule in the River project and update the version of
each River dependency to the version found in env `$VERSION`.

This job is made somewhat easier because Go has a package to easily
parse and write `go.mod` files, so changing values within one is a low
risk operation.

A successful run looks like:

    $ export VERSION=v0.0.14
    $ go run ./internal/cmd/update-submodule-versions/main.go
    go.mod
        set version to v0.0.14 for github.com/riverqueue/river/riverdriver
        set version to v0.0.14 for github.com/riverqueue/river/riverdriver/riverdatabasesql
        set version to v0.0.14 for github.com/riverqueue/river/riverdriver/riverpgxv5
    riverdriver/go.mod
        no changes
    riverdriver/riverdatabasesql/go.mod
        set version to v0.0.14 for github.com/riverqueue/river/riverdriver
    riverdriver/riverpgxv5/go.mod
        set version to v0.0.14 for github.com/riverqueue/river/riverdriver

I figure that more automation is probably possible in the future, but
this is an easy way to give us a fast/reliable way today of changing
versions for a release without accidentally forgetting about one of the
`go.mod`s (which, without the script, is quite easy to do).