Fix data race in rescuer (#120)
Here, the data race described here in the rescuer. It's a little subtle,
but the basic problem is that `defer` statements are LIFO, so what can
happen is that the rescuer can be closed while there's still a log
statement deferred on the stack. Because the log statements feed to
`t.Logf` through a shunt, this can occasionally cause a log line to be
sent to test output after the test has ended, causing the race detector
to detect a race.

The solution is to `defer` the close operation before the log operation.
All the other maintenance services already had this race condition
patched quite some time ago, but the fix didn't make it to the rescuer,
probably due to an outdated branch or the like.

[1] https://github.com/riverqueue/river/pull/116#issuecomment-1858986062