Have GitHub Actions status badge show `master` branch status only (#130)
I just happened to notice while looking at the repo today that the
GitHub Actions status badge reads "failing" if the last Actions run
failed, even if that run was not on the `master` branch. This gives a
somewhat misleading view of the project, implying that the master build
is broken when it's actually not.

Here, scope the badge to the `master` branch only to fix the problem.
The status of non-master branch can easily be viewed from pull requests
or the GitHub Actions runs list.