add JobCancel API to cancel a running job
This adds a new `Client` method `Cancel` which can cancel a running job.
It has these behaviors:

* Jobs which are still scheduled, retryable, or available are
  immediately cancelled and will not be run again.
* Jobs that are already finalized (completed, cancelled, or discarded)
  are not touched.
* Jobs that are actively running are marked for cancellation, and
  cancellation is attempted using a `LISTEN`/`NOTIFY` pubsub message
  which is picked up by the client and producer running that job.

Because Go offers no way to interrupt a running goroutine, actively
running jobs cannot be immediately halted and cancelled, so we can only
cancel the job's context. Once the cancellation signal is received by
the client running the job, any error returned by that job will result
in it being cancelled permanently and not retried. However if the job
returns no error, it will be completed as usual.

In the event the running job finishes executing _before_ the cancellation
signal is received, the job will be completed or retried as usual without
regard to the cancellation. TODO: fix this w/ smarter completion query that
uses metadata value?