add Client to work context, helpers to extract it (#145)
This change adds the `*Client` to an unexported context key within the
work context when the Client is started, making it available to the
context for every job that's worked by that Client.

Additionally, two new helper functions are added to retrieve the Client
from the context within a Worker: `ClientFromContext` and
`ClientFromContextSafely`. The only difference between these is that the
first variant panics if the client isn't found in the context, whereas
the latter returns an error. However if these are used on the context
provided to a Worker's `Work()` method (or one derived from it) they
should never panic/error.