Try to fix release instructions, again
Unfortunately, the release instructions are still wrong. When running
`update-submodule-versions` changes are made to local `go.mod` files,
which need to be pushed and reviewed in a pull request because of the
repo's branch protection.

Here, try to fix the release instructions by wrapping up the updates to
`CHANGELOG.md` and the `go.mod` files into one step, and move it a
little further down the instruction list to make that possible.