Respect job uniqueness when inserting periodic jobs (#168)
As reported in #161, unique jobs combined with periodic jobs doesn't
work as expected. This is because the periodic job enqueueing service
inserts jobs in bulk as an optimization in case many periodic jobs need
inserting, and bulk insertion doesn't support uniqueness.

Here, check for job uniqueness and make sure to use the single job
insertion path when a periodic job is unique. I left the bulk job
insertion path in as well, for use when job uniqueness isn't required.
The optimization still feels like a reasonable trade off to me because
it doesn't make the code that much harder to read, and I expect
non-unique periodic jobs to be the common case.

Fixes #161.