Tee up `go.mod` and `CHANGELOG.md` for version 0.0.17 (#172)
Here, as proposed by the new release instructions in #169, tee up some
changes for 0.0.17, updating `CHANGELOG.md` with the new section and
modifying `go.mod` files to point to the new release.

Unfortunate note no. 1: the `go.mod` files change from 0.0.15 to 0.0.17
instead of starting at 0.0.16 because I accidentally botched the last
release by not pushing the `go.mod` changes. Hopefully this new process
will correct that so it doesn't happen again.

Unfortunate note no. 2: As noted in #169, we expect the build to fail
here because `go.mod` files point to a version that doesn't have a tag
yet.

The new `river validate` is included here (#170) so I'm also going to
try a two-phase release by tagging all submodules except for
`./cmd/river` with 0.0.17 first, then modifying that submodule to point
to the new 0.0.17 in other modules, then do a cut and release for it
separately. Kind of a pain, but as discussed at length before, this
seems to be the least bad of a bunch of suboptimal options.