Specify specific cancel cause on shutdown and use it to determine error logging
Follow up to a discussion [1] in which an additional error log was
causing example tests to fail after uses of `log` were replaced with
`slog` and no longer suppressed.

Here, use `WithCancelCause` to send a specific cancellation error on
shutdown, which can be handled specially for instances where we'd only
want to log an error under unusual circumstances.

[1] https://github.com/riverqueue/river/pull/140#discussion_r1439099623