Small `ci.yml` clean up to remove `BIN_PATH`
A small clean up in `ci.yml` to remove the `BIN_PATH` variable, which
was previously used for sqlc before #82 moved that to a GitHub Action.

It's not a particular problem that `BIN_PATH` is there, but the comment
above it that it's added to `PATH` is no longer true, so probably best
to remove it for accuracy's sake.