leadership election: failing test case to verify #197
This adds thorough test cases around leadership election queries,
including one which demonstrates the issue reported in #197. Not only
does this bug mean that the timestamp would be updated even when a
non-leader attempts to re-elect itself, but also that non leader would
have been tricked into thinking it was still the leader (due to the rows
altered count being >0).