Fix leadership reelection query (#199)
This adds thorough test cases around leadership election queries,
including one which demonstrates the issue reported in #197. Not only
does this bug mean that the timestamp would be updated even when a
non-leader attempts to re-elect itself, but also that non leader would
have been tricked into thinking it was still the leader (due to the rows
altered count being >0).

`EXCLUDED` refers specifically to the params we _attempted_ to insert
before hitting a conflict, so this condition was always evaluating to
true.

Fixes #197. Thank you @joshchoo 🙏🏻