tweak JobRetry to reset ScheduledAt in most cases (#211)
A job which has already completed will have a ScheduledAt that could be
long in the past. Now that we're re-scheduling it, we should update that
to the current time to slot it in alongside other recently-scheduled
jobs and not skip the line. Also, its wait duration can't be calculated
accurately if we don't reset the ScheduledAt (it could show days of wait
time even if it was picked up immediately after being retried).