Minor logging improvements; hopefully to help route out intermittent test
Some very, very small logging improvements that log when the notifier
had started and stopped, like we already have for many other services.
This is an attempt to help diagnose a future CI failure of #213, since
I'm completely unable to reproduce this locally regardless of iteration
count.

Also, augment a log line in the client so it's prefixed with its service
name and contains a `client_id`, which is useful here because it helps
indicate which client ID was elected leader from logs alone.