fix leadership election TTL issue
Due to a glaring oversight in the leader elector code, it actually uses
the same interval for both the internal ticker (for re-electing itself)
and for extending the TTL in the database. These timers and queries are
of course not instantaneous, which means a perfectly functional leader
can fail to re-elect itself cleanly.

Extensive test coverage on the elector to come in subsequent PRs.

Fixes #216.