Prepare `v0.0.21` release
Prepare a new release for `v0.0.21`, with the most important change
contained being #217, which fixes a bug in the leadership election code.

I know I said I'd switch to minor version releases after `v0.0.20`, but
since this version is largely a hotfix with a small change, I did
another patch.