Reuse timer on `DebouncedChan` + put in some comments + continous test
I wanted to understand a little more clearly how `DebouncedChan` works,
so I went in to look for a few refactoring opportunities. There was a
TODO in the file to potentially reuse the internal timer so it doesn't
need to be reallocated every debounce cycle, so here we implement that.

To make sure that the reused timer is really working we also add a new
"continuous" test that checks that the channel will in fact fire many
times in a row. We also go through and add a few comments and rename a
couple things for clarity, and bring in a `TestMain` wrapper so that the
package tests are checked for goroutine leaks.