Upgrade golangci-lint GitHub Action go `v5` and linter to 1.56 (#223)
A small one to upgrade:

* golangci-lint GitHub Action from `v4` to `v5`
* golangci-lint from 1.55 to 1.56

A small change, but I upgraded locally at some point, and it's causing a
build issue in #212. The `tparallel` lint previously had a bug where
it'd want you to put a `t.Parallel()` on test cases that use `t.Setenv()`,
even though use of `t.Parallel()` with `t.Setenv()` is disallowed by Go's
test framework, so you'd have to mark those test cases with `nolint`.

That bug's been fixed now, and now the linter detects that the `nolint`
is no longer necessary, and automatically strips it out with my upgraded
golangci-lint version, causing a failure when I push to CI.

The `v5` GitHub Action is also somewhat nice because it stops using a
deprecated NodeJS version, so produces fewer warnings in the log.