Fix intermittent test problem in `rivertest` + more transactional use
This one's aimed at fixing #225, in which I think what was happening is
that because we were inserting a test job on a non-transactional pool
with a state of `scheduled`, in some cases the queue maintainer's
scheduler service was actually changing it back to `available` before
the test case would check it with a `RequireInserted`.

As I was moving through the test suite, I realized that we could be
making a lot more use of `InsertTx` variants pretty much everywhere
since we were mostly testing the transactional variables of
`RequireInserted`, so that I changed those over.

Then I realize that even when we were testing the non-transactional
variants, we didn't actually need a started client to do so, so I took
that code out. Then I realized that since we're not starting the client,
we didn't actually need to specify `Queues` or `Workers` (I believe the
idea of an insert-only came about only after I wrote these tests
originally). And if we didn't need to specify `Workers`, then we didn't
even need the worker structs, so I got rid of those too.

Overall, I think we simplify things quite a bit, and because the client
doesn't need to be started, probably reduces the possibility of future
intermittency problems like we observed in #225.