Drop sqlc overrides for `JobState`
This is another small one, but I noticed while doing driver work that
since we put a layer of abstraction between sqlc and any public-facing
APIs, the `JobState*` state constants in sqlc are not really used
anymore. Even the `JobState` type is used to infrequently that golfing
"River" off the front also does very little.

Here, I propose that we just get rid of the sqlc overrides. They're not
a big deal, but not having them means less sqlc configuration that needs
to be shared around between YAML files, and less opportunity for error
there.