Move `JobCompleteTx` to its own `.go` file
Another small one, but I always have a hard time remembering where
`JobCompleteTx` lives because it doesn't feel intuitive to me that it
should belong in `job.go` with the `Job` definition.

Here, break it out into its own separate file with a clear-as-day name
associated with it. Also give it a basic test suite that pairs with the
new file. This'll be somewhat duplicative with some cases in
`client_test.go`, but that shouldn't be harmful.