Proof of concept: River operating with pgx simple protocol
We had an issue opened a little while back (#171) wherein it turned out
that River couldn't operate with pgx's simple protocol turned on, which
appeared to be because the simple protocol doesn't know anything about a
target Postgres tuple's makeup and therefore had to guess how to encode
parameters based on input types, which was leading `[]byte` intended for
`jsonb`s to be binary encoded, which isn't allowed.

We closed that issue after I pointed out that PgBouncer now supports
prepared statement pooling, and therefore simple protocol is no longer
necessary.

We recently got another comment [1] where someone had tried to use River
with Supavisor on Supabase, which also does not supported prepared
statements, leading to lots of errors. Normally, pgx's simple protocol
could be used to address this, but as we saw in #171, River doesn't
support that.

Here, a proof of concept demonstrating how River could support simple
protocol by modeling `jsonb` fields in sqlc as strings, which prevents
pgx from encoding to binary with simple procotol, and we get to a point
where all driver tests are passing.

I'm not sure if we should action on this one or not. Having to fall back
to `string` fields a bit gross, although I don't *think* carries
significant performance penalty because it's basically the same
underlying structure as far as Go is concerned.

[1] https://github.com/riverqueue/river/issues/205#issuecomment-1963001669