`Rescuer` -> `JobRescuer` + `Scheduler` -> `JobScheduler` (#233)
Another small one, but it's triggering my OCD in that for services
inside the queue maintainer, the job cleaner is called `JobCleaner`, the
periodic enqueuer is called `PeriodicJobEnqueuer`, but the rescuer is
only `Rescuer` (without a "job" prefix) and scheduler is `Scheduler`.

Here, rename the rescuer and scheduler so they pick up a prefix and
become `JobRescuer` and `JobScheduler`. I think this is nominally
beneficial anyway because it helps telegraph that these are services
that are specific to jobs (unlike say `Reindexer` for example, which is
not related to jobs specifically, and therefore unprefixed).