Don't disable sleep by default from `riverinternaltest.BaseServiceArchetype` (#235)
The `riverinternaltest.BaseServiceArchetype` test helper has up until
now return an archetype with sleep disabled, with the idea that
deactivating it by default derisked the possibility that tests would
accidentally introduce sleep invocations that'd make them slow.

(Note this only works for invocations of `BaseService.CancellableSleep`
which a lot of River doesn't use.)

This made more sense when the main use of the archetype was in in the
queue maintenance services, which all start up with a short jittered
sleep so that they don't all go for the database at the same time.

But since then, the use of archetype has become more widespread, and
disabling sleep probably doesn't make sense in most situations,
especially when testing at the client level. It'd have the effect of
making all the queue maintenance services run more aggressively than
they would by default, which isn't desirable, and may even lead to
contention problems.

Here, add a `WithSleepDisabled()` helper to archetype that lets sleep be
disabled easily in tests that want to do that (and add invocations of it
to each of the queue maintainer tests), but keep sleep on everywhere
else.